Krewlancer-main/Krewlancer-main/legacy_backend_archive/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
Flask-Limiter==3.8.0
Flask-SQLAlchemy==3.1.1
Werkzeug==3.0.0
python-dotenv==1.0.0
razorpay==2.0.0
Flask-Mail==0.9.1
authlib
requests
pyOpenSSL
PyMySQL==1.1.2
pymongo==4.10.1
gunicorn==21.2.0